Advanced logging with dynamic config

+++ src/support/feature.js
@@ -9,39 +9,68 @@ const onSubmit = require("./modals");
 
 class SupportFeature {
 
-    constructor(bot, config) {
+    static tables = { 
 
+        /**
+         * @type { TicketTypeTable }
+         */
+        ticketTypes: null, 
+        
+        /**
+         * @type { TicketStatusTable }
+         */
+        ticketStatus: null,
+        
+        /**
+         * @type { TicketTable }
+         */
+        tickets: null, 
+        
+        /**
+         * @type { TicketMessagesTable }
+         */
+        transcript: null 
+
+    }
+
+    constructor(bot) {
+
+        /**
+         * @type { import("../bot") }
+         */
         this.bot = bot
-        this.config = config
-
-        Object.entries(config).forEach(([key, value]) => this[key] = value)
 
         commands.forEach(([ cmdData, cmdPerms ]) => this.bot.commands.add(cmdData, cmdPerms))
 
-        this.transcriptChannel = null
+        this.transcriptChannels = {}
+        this.ticketCategorys = {}
 
+        this.database.addTable("ticketStatus", new TicketStatusTable(this.database))
+        this.database.addTable("ticketTypes", new TicketTypeTable(this.database))
+        this.database.addTable("tickets", new TicketTable(this.database))
+        
+        this.database.addTable("transcript", new TicketMessagesTable(this.database))
+        
         bot.on('ready', () => this.onReady())
+
+    }
+
+    get database() {
+
+        return this.bot.database;
 
     }
 
     async onReady() {
 
-        this.bot.database.tickets = new TicketTable(this.bot.database)
-        this.bot.database.ticketTypes = new TicketTypeTable(this.bot.database)
-        this.bot.database.ticketStatus = new TicketStatusTable(this.bot.database)
-        this.bot.database.transcript = new TicketMessagesTable(this.bot.database)
+        this.transcriber = new TicketTranscriber(this.database.transcript)
 
-        await this.bot.database.tickets.connect()
-        await this.bot.database.ticketTypes.connect()
-        await this.bot.database.ticketStatus.connect()
-        await this.bot.database.transcript.connect()
+        const channelConfigs = this.database.guildEntrys.cache.get({ type: { name: "tickets_transcript_channel" } })
+        await Promise.all(channelConfigs.map(entry => this.setTranscriptChannel(entry.guild_id, entry.value)))
 
-        this.transcriber = new TicketTranscriber(this.bot.database.transcript)
-
-        if (this.transcriptChannelId) {
-            this.transcriptChannel = await this.bot.channels.fetch(this.transcriptChannelId)
-            console.log("Transcript channel found and on standby!")
-        }
+        this.database.guildEntrys.cache.on('push', config => this.onConfigCreate(config))
+        this.database.guildEntrys.cache.on('update', config => this.onConfigCreate(config))
+        this.database.guildEntrys.cache.on('remove', config => this.onConfigRemove(config))
 
         this.addEventHandlers()
 
@@ -50,13 +79,117 @@ class SupportFeature {
         this.bot.on('messageDeleteBulk', messages => this.onMessageDeleteBulk(messages))
         this.bot.on('messageUpdate', (oldMessage, newMessage) => this.onMessageUpdate(oldMessage, newMessage))
 
-        this.bot.on('channelDelete', channel => this.onChannelDelete(channel))
+        this.bot.on('scrimsChannelDelete', channel => this.onChannelDelete(channel))
+
+    }
+
+    async onConfigCreate(config) {
+
+        if (config.type.name == "tickets_transcript_channel") {
+
+            await this.setTranscriptChannel(config.guild_id, config.value)
+
+        }
+
+        if (config.type.name == "tickets_report_category") {
+
+            await this.setTicketsCategory(config.guild_id, config.value, 'report')
+
+        }
+
+        if (config.type.name == "tickets_support_category") {
+
+            await this.setTicketsCategory(config.guild_id, config.value, 'support')
+
+        }
+
+    }
+
+    async onConfigRemove(config) {
+
+        if (config.type.name == "tickets_transcript_channel") {
+
+            delete this.transcriptChannels[config.guild_id]
+            this.logError(`Transcript channel unconfigured!`, { guild_id: config.guild_id })
+
+        }
+
+        if (config.type.name == "tickets_report_category") {
+
+            delete this.ticketCategorys[config.guild_id]?.report
+            this.logError(`Ticket report category unconfigured!`, { guild_id: config.guild_id })
+
+        }
+
+        if (config.type.name == "tickets_support_category") {
+
+            delete this.ticketCategorys[config.guild_id]?.support
+            this.logError(`Ticket support category unconfigured!`, { guild_id: config.guild_id })
+
+        }
+
+    }
+
+    getTicketCategory(guildId, typeName) {
+
+        if (!(guildId in this.ticketCategorys)) return null;
+
+        return this.ticketCategorys[guildId][typeName] ?? null;
+
+    }
+
+    getTranscriptChannel(guildId) {
+
+        return this.transcriptChannels[guildId] ?? null;
+
+    }
+
+    async setTranscriptChannel(guildId, channelId) {
+
+        const channel = await this.bot.channels.fetch(channelId)
+            .catch(error => this.logError(`Fetching tickets transcript channel failed!`, { guild_id: guildId, error }))
+
+        if (channel) {
+
+            this.logSuccess(`Transcript channel set as **${channel.name}**.`, { guild_id: guildId })
+            this.transcriptChannels[guildId] = channel
+
+        }
+
+    }
+
+    async setTicketsCategory(guildId, channelId, typeName) {
+
+        const channel = await this.bot.channels.fetch(channelId)
+            .catch(error => this.logError(`Fetching ${typeName} ticket category failed!`, { guild_id: guildId, error }))
+
+        if (channel) {
+
+            this.logSuccess(`The category for ${typeName} tickets set as **${channel.name}**.`, { guild_id: guildId })
+            
+            if (!(guildId in this.ticketCategorys)) this.ticketCategorys[guildId] = {}
+            this.ticketCategorys[guildId][typeName] = channel
+
+        }
+
+    }
+
+    logError(msg, context) {
+
+        if (context.error) console.error(`${msg} Reason: ${context.error}`)
+        this.database.ipc.notify(`support_error`, { msg, ...context })
+
+    }
+
+    logSuccess(msg, context) {
+
+        this.database.ipc.notify(`support_success`, { msg, ...context })
 
     }
 
     async onMessageCreate(message) {
 
-        const ticket = this.bot.database.tickets.cache.get({ channel_id: message.channel.id })[0]
+        const ticket = this.database.tickets.cache.get({ channel_id: message.channel.id })[0]
         if (!ticket) return false;
 
         if (message.author.id == this.bot.user.id) return false;
@@ -72,7 +205,7 @@ class SupportFeature {
 
         }
 
-        await this.bot.database.transcript.create(ticketMessage)
+        await this.database.transcript.create(ticketMessage)
             .catch(error => console.error(`Unable to log support ticket message because of ${error}`, ticketMessage))
 
     }
@@ -86,10 +219,10 @@ class SupportFeature {
 
     async onMessageDelete(message) {
 
-        const ticket = this.bot.database.tickets.cache.get({ channel_id: message.channel.id })[0]
+        const ticket = this.database.tickets.cache.get({ channel_id: message.channel.id })[0]
         if (!ticket) return false;
 
-        await this.bot.database.transcript.update({ id_ticket: ticket.id_ticket, message_id: message.id }, { deleted: Math.round(Date.now() / 1000) })
+        await this.database.transcript.update({ id_ticket: ticket.id_ticket, message_id: message.id }, { deleted: Math.round(Date.now() / 1000) })
             .catch(error => console.error(`Unable to log support ticket message deletion because of ${error}`, ticket))
 
     }
@@ -115,9 +248,16 @@ class SupportFeature {
 
         }
 
-        const existing = await interaction.client.database.tickets.get({ type: { name: typeName }, id_user: interaction.scrimsUser.id_user, status: { name: "open" } })
-        if (existing.length > 0)
-            return interaction.reply(ScrimsMessageBuilder.errorMessage(`Already Created`, `You already have a ticket of this type open (<#${existing[0].channel_id}>)!`)).then(() => false);
+        const existing = await this.database.tickets.get({ type: { name: typeName }, id_user: interaction.scrimsUser.id_user, status: { name: "open" } })
+        if (existing.length > 0) {
+
+            const channel = await this.bot.channels.fetch(existing[0].channel_id).catch(() => null)
+            if (channel) return interaction.reply(ScrimsMessageBuilder.errorMessage(`Already Created`, `You already have a ticket of this type open (${channel})!`)).then(() => false);
+
+            // Ticket is open, but the channel does not exist
+            await this.closeTicket({ guild: interaction.guild }, existing[0], null)
+        
+        }
 
         return true;
 
@@ -125,17 +265,39 @@ class SupportFeature {
 
     async onChannelDelete(channel) {
 
-        const tickets = await this.bot.database.tickets.get({ channel_id: channel.id, status: { name: "open" } }).catch(console.error)
-        await Promise.all(tickets.map(ticket => this.closeTicket(channel, ticket))).catch(console.error)
+        const transcriptChannel = this.getTranscriptChannel(channel.guild.id)
+        if (transcriptChannel?.id === channel.id) {
+
+            delete this.transcriptChannels[channel.guild.id]
+            this.logError(`Deleted the transcript channel!`, { guild_id: channel.guild.id, executor_id: channel?.executor?.id })
+
+        }
+
+        const categorys = this.ticketCategorys[channel.guild.id] ?? {}
+        const hits = Object.entries(categorys).filter(([_, category]) => category.id === channel.id).map(([typeName, _]) => typeName)
+        hits.forEach(typeName => delete categorys[typeName])
+
+        this.logError(
+            `Deleted the ${hits.join('and')} ticket category${(hits.length > 1) ? 's' : ''}!`, 
+            { guild_id: channel.guild.id, executor_id: channel?.executor?.id }
+        )
+
+        const types = [ 'tickets_transcript_channel', 'tickets_report_category', 'tickets_support_category' ]
+        await Promise.all(types.map(name => this.database.guildEntrys.remove({ guild_id: channel.guild.id, type: { name }, value: channel.id }))).catch(console.error)
+
+        const tickets = await this.database.tickets.get({ channel_id: channel.id, status: { name: "open" } }).catch(console.error)
+        await Promise.all(tickets.map(ticket => this.closeTicket(channel, ticket, channel?.executor))).catch(console.error)
 
     }
 
-    async closeTicket(channel, ticket) {
+    async closeTicket(channel, ticket, executor) {
 
+        this.logError(`Closed a ticket.`, { guild_id: channel.guild.id, ticket, executor_id: (executor?.id ?? null) })
         await this.transcriber.send(channel.guild, ticket)
 
-        await this.bot.database.tickets.update({ id_ticket: ticket.id_ticket }, { status: { name: "deleted" } })
-        await channel.delete().catch(() => { /* Channel could already be deleted. */ })
+        await this.database.tickets.update({ id_ticket: ticket.id_ticket }, { status: { name: "deleted" } })
+        
+        if (typeof channel.delete === "function") await channel.delete().catch(() => { /* Channel could already be deleted. */ })
 
     }
